fix: no more overlaps

+++ iarbre_data/management/commands/c02_init_grid.py
@@ -75,8 +75,8 @@ def create_hexs_for_city(
     xmax = np.ceil(xmax / hex_width) * hex_width
     ymax = np.ceil(ymax / hex_height) * hex_height
 
-    cols = np.arange(np.floor(xmin), np.ceil(xmax), 3 * unit)
-    rows = np.arange(np.floor(ymin) / a, np.ceil(ymax) / a, unit)
+    cols = np.arange(xmin, xmax, 3 * unit)
+    rows = np.arange(ymin / a, ymax / a, unit)
     tiles = []
     for x, (i, y) in tqdm(itertools.product(cols, enumerate(rows))):
         # Rows are not aligned
@@ -128,6 +128,12 @@ class Command(BaseCommand):
         parser.add_argument(
             "--grid-type", type=int, default=1, help="Hexagonal (1) or square (2) grid."
         )
+        parser.add_argument(
+            "--clean_outside",
+            type=bool,
+            default=True,
+            help="Detele tiles outside of the city boundaries of the selection or not.",
+        )
 
     @staticmethod
     def _remove_duplicates():
@@ -152,6 +158,7 @@ class Command(BaseCommand):
         insee_code_city = options["insee_code_city"]
         grid_size = options["grid_size"]
         grid_type = options["grid_type"]
+        clean_outside = options["clean_outside"]
         if grid_type not in [1, 2]:
             raise ValueError("Grid type should be either 1 (hexagonal) or 2 (square).")
         if insee_code_city is not None:  # Perform selection only for a city
@@ -195,17 +202,17 @@ class Command(BaseCommand):
 
         print("Removing duplicates...")
         self._remove_duplicates()
-
-        # Clean useless tiles
-        city_union_geom = selected_city.geometry.unary_union
-        print("Deleting tiles out of the cities")
-        total_records = Tile.objects.all().count()
-        for start in tqdm(range(0, total_records, batch_size * 10)):
-            batch_ids = Tile.objects.all()[start : start + batch_size * 10].values_list(
-                "id", flat=True
-            )
-            with transaction.atomic():
-                Tile.objects.filter(id__in=batch_ids).exclude(
-                    geometry__intersects=city_union_geom.wkt
-                ).delete()
-        logger.info(f"Deleted {total_records} tiles")
+        if clean_outside:
+            # Clean useless tiles
+            city_union_geom = selected_city.geometry.unary_union
+            print("Deleting tiles out of the cities")
+            total_records = Tile.objects.all().count()
+            for start in tqdm(range(0, total_records, batch_size * 10)):
+                batch_ids = Tile.objects.all()[
+                    start : start + batch_size * 10
+                ].values_list("id", flat=True)
+                with transaction.atomic():
+                    Tile.objects.filter(id__in=batch_ids).exclude(
+                        geometry__intersects=city_union_geom.wkt
+                    ).delete()
+            logger.info(f"Deleted {total_records} tiles")
